refactor(api): isolate server mutations
